MaltaeC/requirements.txt
Django==1.11.5
redis==2.10.3
django-redis==4.3.0
gunicorn==19.3.0
psycopg2
chardet==3.0.4
mysqlclient
django-configurations==0.8
django-secure==1.0
django-cache-url==0.8.0
dj-database-url==0.3.0

django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.2.0

django-model-utils==2.2

Pillow==2.5.3

django-allauth==0.18.0

unicode-slugify==0.1.1
django-autoslug==1.7.2

django-avatar==2.0
coverage==3.7.1
Sphinx
django-debug-toolbar==1.2.1
python-memcached
django-localflavor
stripe
pycrypto
passlib
dj-static
whitenoise
pyyaml
ua-parser
user-agents
django-user-agents
